fix: branch conditions
