+++ packages/react-ui/src/app/builder/step-settings/branch-settings/branch-condition-group.tsx
@@ -1,162 +1,71 @@
-import { t } from 'i18next';
-import { Trash } from 'lucide-react';
-import { useFormContext } from 'react-hook-form';
+import React from "react";
+import { useFieldArray, useFormContext } from "react-hook-form";
+import { BranchAction } from "@activepieces/shared";
+import { HorizontalSeparatorWithText } from "@/components/ui/seperator";
+import { t } from "i18next";
+import { BranchSingleCondition } from "@/app/builder/step-settings/branch-settings/branch-single-condition";
+import { BranchConditionToolbar } from "@/app/builder/step-settings/branch-settings/branch-condition-toolbar";
 
-import { SearchableSelect } from '@/components/custom/searchable-select';
-import { Button } from '@/components/ui/button';
-import { FormField, FormItem, FormMessage } from '@/components/ui/form';
-import { Label } from '@/components/ui/label';
-import { Switch } from '@/components/ui/switch';
-import { cn } from '@/lib/utils';
-import {
-  BranchOperator,
-  textConditions,
-  singleValueConditions,
-  BranchAction,
-} from '@activepieces/shared';
 
-import { TextInputWithMentions } from '../../piece-properties/text-input-with-mentions';
-
-const textToBranchOperation: Record<BranchOperator, string> = {
-  [BranchOperator.TEXT_CONTAINS]: t('(Text) Contains'),
-  [BranchOperator.TEXT_DOES_NOT_CONTAIN]: t('(Text) Does not contain'),
-  [BranchOperator.TEXT_EXACTLY_MATCHES]: t('(Text) Exactly matches'),
-  [BranchOperator.TEXT_DOES_NOT_EXACTLY_MATCH]: t(
-    '(Text) Does not exactly match',
-  ),
-  [BranchOperator.TEXT_STARTS_WITH]: t('(Text) Starts with'),
-  [BranchOperator.TEXT_DOES_NOT_START_WITH]: t('(Text) Does not start with'),
-  [BranchOperator.TEXT_ENDS_WITH]: t('(Text) Ends with'),
-  [BranchOperator.TEXT_DOES_NOT_END_WITH]: t('(Text) Does not end with'),
-  [BranchOperator.NUMBER_IS_GREATER_THAN]: t('(Number) Is greater than'),
-  [BranchOperator.NUMBER_IS_LESS_THAN]: t('(Number) Is less than'),
-  [BranchOperator.NUMBER_IS_EQUAL_TO]: t('(Number) Is equal to'),
-  [BranchOperator.BOOLEAN_IS_TRUE]: t('(Boolean) Is true'),
-  [BranchOperator.BOOLEAN_IS_FALSE]: t('(Boolean) Is false'),
-  [BranchOperator.EXISTS]: t('Exists'),
-  [BranchOperator.DOES_NOT_EXIST]: t('Does not exist'),
-};
-const operationOptions = Object.keys(textToBranchOperation).map((operator) => {
-  return {
-    label: textToBranchOperation[operator as BranchOperator],
-    value: operator,
+type BranchConditionGroupProps = {
+    readonly: boolean;
+    groupIndex: number;
+    onAnd: () => void;
+    onOr: () => void;
+    numberOfGroups: number;
+    handleDelete: (conditionIndex: number) => void;
   };
-});
-
-type BranchSingleConditionProps = {
-  showDelete: boolean;
-  groupIndex: number;
-  conditionIndex: number;
-  readonly: boolean;
-  deleteClick: () => void;
-};
-
-const BranchSingleCondition = ({
-  deleteClick,
-  groupIndex,
-  conditionIndex,
-  showDelete,
-  readonly,
-}: BranchSingleConditionProps) => {
-  const form = useFormContext<BranchAction>();
-
-  const condition =
-    form.getValues().settings.conditions[groupIndex][conditionIndex];
-  const isTextCondition =
-    condition.operator && textConditions.includes(condition?.operator);
-  const isSingleValueCondition =
-    condition.operator && singleValueConditions.includes(condition?.operator);
-  return (
-    <>
-      <div
-        className={cn('grid gap-2', {
-          'grid-cols-2': isSingleValueCondition,
-          'grid-cols-3': !isSingleValueCondition,
-        })}
-      >
-        <FormField
-          name={`settings.conditions.${groupIndex}.${conditionIndex}.firstValue`}
-          control={form.control}
-          render={({ field }) => (
-            <FormItem>
-              <TextInputWithMentions
-                disabled={readonly}
-                placeholder={t('First value')}
-                onChange={field.onChange}
-                initialValue={field.value}
-              ></TextInputWithMentions>
-              <FormMessage />
-            </FormItem>
+  
+  const BranchConditionGroup = React.memo(
+    ({ readonly, groupIndex, onAnd, onOr, handleDelete, numberOfGroups }: BranchConditionGroupProps) => {
+      const form = useFormContext<BranchAction>();
+      const { fields } = useFieldArray({
+        control: form.control,
+        name: `settings.conditions.${groupIndex}`,
+      });
+  
+      return (
+        <div className="flex flex-col gap-4">
+          {groupIndex > 0 && (
+            <HorizontalSeparatorWithText className="my-2">
+              {t('OR')}
+            </HorizontalSeparatorWithText>
           )}
-        />
-        <FormField
-          name={`settings.conditions.${groupIndex}.${conditionIndex}.operator`}
-          control={form.control}
-          render={({ field }) => (
-            <FormItem>
-              <SearchableSelect
-                disabled={readonly}
-                value={field.value}
-                options={operationOptions}
-                placeholder={''}
-                onChange={(e) => field.onChange(e)}
+          {fields.length === 0 && (
+            <BranchConditionToolbar
+              readonly={readonly}
+              key={`toolbar-${groupIndex}`}
+              onAnd={onAnd}
+              onOr={onOr}
+              showOr={true}
+              showAnd={true}
+            />
+          )}
+          {fields.map((condition, conditionIndex) => (
+            <React.Fragment key={condition.id}>
+              {conditionIndex > 0 && <div>{t('And If')}</div>}
+              <BranchSingleCondition
+                groupIndex={groupIndex}
+                readonly={readonly}
+                conditionIndex={conditionIndex}
+                deleteClick={() => handleDelete(conditionIndex)}
+                showDelete={numberOfGroups !== 1 || fields.length !== 1}
               />
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        {!isSingleValueCondition && (
-          <FormField
-            name={`settings.conditions.${groupIndex}.${conditionIndex}.secondValue`}
-            control={form.control}
-            render={({ field }) => (
-              <FormItem>
-                <TextInputWithMentions
-                  placeholder={t('Second value')}
-                  disabled={readonly}
-                  initialValue={field.value}
-                  onChange={field.onChange}
-                ></TextInputWithMentions>
-                <FormMessage />
-              </FormItem>
-            )}
+            </React.Fragment>
+          ))}
+          <BranchConditionToolbar
+            onAnd={onAnd}
+            onOr={onOr}
+            readonly={readonly}
+            showOr={true}
+            showAnd={true}
           />
-        )}
-      </div>
-
-      <div className="flex justify-start items-center gap-2 mt-2">
-        {isTextCondition && (
-          <FormField
-            name={`settings.conditions.${groupIndex}.${conditionIndex}.caseSensitive`}
-            control={form.control}
-            render={({ field }) => (
-              <FormItem>
-                <div className="flex items-center gap-2 p-1">
-                  <Switch
-                    disabled={readonly}
-                    id="case-sensitive"
-                    checked={field.value}
-                    onCheckedChange={(e) => field.onChange(e)}
-                  />
-                  <Label htmlFor="case-sensitive">{t('Case sensitive')}</Label>
-                </div>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-        )}
-        <div className="flex-grow"></div>
-        <div>
-          {showDelete && (
-            <Button variant={'basic'} size={'sm'} onClick={deleteClick}>
-              <Trash className="w-4 h-4"></Trash> {t('Remove')}
-            </Button>
-          )}
         </div>
-      </div>
-    </>
+      );
+    }
   );
-};
-
-BranchSingleCondition.displayName = 'BranchSingleCondition';
-export { BranchSingleCondition };
+  
+  BranchConditionGroup.displayName = 'ConditionGroup';
+  
+  export { BranchConditionGroup };
+
