+++ packages/react-ui/src/app/builder/step-settings/branch-settings/index.tsx
@@ -1,16 +1,14 @@
 import { t } from 'i18next';
 import React from 'react';
-import { useFormContext } from 'react-hook-form';
+import { useFormContext, useFieldArray } from 'react-hook-form';
 
-import { HorizontalSeparatorWithText } from '@/components/ui/seperator';
 import {
   BranchAction,
   BranchOperator,
   ValidBranchCondition,
 } from '@activepieces/shared';
+import { BranchConditionGroup } from '@/app/builder/step-settings/branch-settings/branch-condition-group';
 
-import { BranchSingleCondition } from './branch-condition-group';
-import { BranchConditionToolbar } from './branch-condition-toolbar';
 
 const emptyCondition: ValidBranchCondition = {
   firstValue: '',
@@ -25,6 +23,10 @@ type BranchSettingsProps = {
 
 const BranchSettings = React.memo(({ readonly }: BranchSettingsProps) => {
   const form = useFormContext<BranchAction>();
+  const { fields, append, remove, update } = useFieldArray({
+    control: form.control,
+    name: 'settings.conditions',
+  });
 
   const handleDelete = (groupIndex: number, conditionIndex: number) => {
     const conditions = form.getValues().settings.conditions;
@@ -32,81 +34,41 @@ const BranchSettings = React.memo(({ readonly }: BranchSettingsProps) => {
     const isSingleGroup = conditions.length === 1;
     const isSingleConditionInGroup = newConditionsGroup.length === 1;
 
-    let newConditions;
-
     if (isSingleGroup && isSingleConditionInGroup) {
-      newConditions = [[emptyCondition]];
+      update(groupIndex, [emptyCondition]);
     } else if (isSingleConditionInGroup) {
-      newConditions = conditions.filter((_, index) => index !== groupIndex);
+      remove(groupIndex);
     } else {
       newConditionsGroup.splice(conditionIndex, 1);
-      newConditions = [...conditions];
-      newConditions[groupIndex] = newConditionsGroup;
+      update(groupIndex, newConditionsGroup);
     }
-    form.setValue('settings.conditions', newConditions, {
-      shouldValidate: true,
-    });
   };
+
   const handleAnd = (groupIndex: number) => {
     const conditions = form.getValues().settings.conditions;
     conditions[groupIndex] = [...conditions[groupIndex], emptyCondition];
-    form.setValue('settings.conditions', conditions, { shouldValidate: true });
+    update(groupIndex, conditions[groupIndex]);
   };
 
   const handleOr = () => {
-    const conditions = form.getValues().settings.conditions;
-    conditions.push([emptyCondition]);
-    form.setValue('settings.conditions', conditions, { shouldValidate: true });
+    append([]);
+    handleAnd(fields.length - 1);
   };
 
-  const conditions = form.getValues().settings.conditions;
   return (
     <div className="flex flex-col gap-4" onSubmit={(e) => e.preventDefault()}>
       <div className="text-md">{t('Continue If')}</div>
-      {conditions.map((fieldGroup, groupIndex) => {
-        return (
-          <div className="flex flex-col gap-4" key={`group-${groupIndex}`}>
-            {groupIndex > 0 && (
-              <HorizontalSeparatorWithText className="my-2">
-                {t('OR')}
-              </HorizontalSeparatorWithText>
-            )}
-            {fieldGroup.length === 0 && (
-              <BranchConditionToolbar
-                readonly={readonly}
-                key={`toolbar-${groupIndex}`}
-                onAnd={() => handleAnd(groupIndex)}
-                onOr={() => handleOr()}
-                showOr={groupIndex === conditions.length - 1}
-                showAnd={true}
-              ></BranchConditionToolbar>
-            )}
-            {fieldGroup.map((condition, conditionIndex) => (
-              <React.Fragment
-                key={`condition-${groupIndex}-${conditionIndex}-${condition.operator}`}
-              >
-                {conditionIndex > 0 && <div>{t('And If')}</div>}
-                <BranchSingleCondition
-                  groupIndex={groupIndex}
-                  readonly={readonly}
-                  conditionIndex={conditionIndex}
-                  deleteClick={() => handleDelete(groupIndex, conditionIndex)}
-                  showDelete={
-                    conditions.length !== 1 || fieldGroup.length !== 1
-                  }
-                ></BranchSingleCondition>
-              </React.Fragment>
-            ))}
-            <BranchConditionToolbar
-              onAnd={() => handleAnd(groupIndex)}
-              onOr={() => handleOr()}
-              readonly={readonly}
-              showOr={groupIndex === conditions.length - 1}
-              showAnd={true}
-            ></BranchConditionToolbar>
-          </div>
-        );
-      })}
+      {fields.map((fieldGroup, groupIndex) => (
+        <BranchConditionGroup
+          key={fieldGroup.id}
+          readonly={readonly}
+          numberOfGroups={fields.length}
+          groupIndex={groupIndex}
+          onAnd={() => handleAnd(groupIndex)}
+          onOr={handleOr}
+          handleDelete={(conditionIndex: number) => handleDelete(groupIndex, conditionIndex)}
+        />
+      ))}
     </div>
   );
 });
